finished making suggestions
