--- examples/Q3b_variational-classifier-iris.py
@@ -1,188 +1,0 @@
-"""Variational quantum classifier.
-
-In this demo we implement a variational classifier inspired by
-Schuld et al. 2018 (arXiv:1804.00633).
-"""
-
-import pennylane as qml
-from pennylane import numpy as np
-from pennylane.optimize import NesterovMomentumOptimizer
-
-dev = qml.device('default.qubit', wires=2)
-
-
-def get_angles(x):
-    """Compute rotation angles for state preparation routine.
-
-    Args:
-        x: original input vector
-
-    Returns:
-        array[float]: feature vector of angles
-    """
-
-    beta0 = 2 * np.arcsin(np.sqrt(x[1]) ** 2 / np.sqrt(x[0] ** 2 + x[1] ** 2 + 1e-12) )
-    beta1 = 2 * np.arcsin(np.sqrt(x[3]) ** 2 / np.sqrt(x[2] ** 2 + x[3] ** 2 + 1e-12) )
-    beta2 = 2 * np.arcsin(np.sqrt(x[2] ** 2 + x[3] ** 2) / np.sqrt(x[0] ** 2 + x[1] ** 2 + x[2] ** 2 + x[3] ** 2))
-
-    return np.array([beta2, -beta1 / 2, beta1 / 2, -beta0 / 2, beta0 / 2])
-
-
-def statepreparation(a):
-
-    """Quantum circuit to encode a positive, normalized 4-d vector
-    into the amplitudes of 2 qubits.
-
-    Args:
-        a: feature vector of rotation angles
-    """
-
-    # encode absolute values
-    qml.RY(a[0], wires=0)
-
-    qml.CNOT(wires=[0, 1])
-    qml.RY(a[1], wires=1)
-    qml.CNOT(wires=[0, 1])
-    qml.RY(a[2], wires=1)
-
-    qml.PauliX(wires=0)
-    qml.CNOT(wires=[0, 1])
-    qml.RY(a[3], wires=1)
-    qml.CNOT(wires=[0, 1])
-    qml.RY(a[4], wires=1)
-    qml.PauliX(wires=0)
-
-
-def layer(W):
-    """ Single layer of the variational classifier.
-
-    Args:
-        W (array[float]): 2-d array of variables for one layer
-    """
-
-    qml.Rot(W[0, 0], W[0, 1], W[0, 2], wires=0)
-    qml.Rot(W[1, 0], W[1, 1], W[1, 2], wires=1)
-
-    qml.CNOT(wires=[0, 1])
-    
-
-@qml.qnode(dev)
-def circuit(weights, angles=None):
-    """The circuit of the variational classifier."""
-
-    statepreparation(angles)
-    
-    for W in weights:
-        layer(W)
-
-    return qml.expval.PauliZ(0)
-
-
-def variational_classifier(var, angles=None):
-    """The variational classifier."""
-
-    weights = var[0]
-    bias = var[1]
-
-    return circuit(weights, angles=angles) + bias
-
-
-def square_loss(labels, predictions):
-    """ Square loss function
-
-    Args:
-        labels (array[float]): 1-d array of labels
-        predictions (array[float]): 1-d array of predictions
-    Returns:
-        float: square loss
-    """
-    loss = 0
-    for l, p in zip(labels, predictions):
-        loss = loss + (l - p) ** 2
-    loss = loss / len(labels)
-
-    return loss
-
-
-def accuracy(labels, predictions):
-    """ Share of equal labels and predictions
-
-    Args:
-        labels (array[float]): 1-d array of labels
-        predictions (array[float]): 1-d array of predictions
-    Returns:
-        float: accuracy
-    """
-
-    loss = 0
-    for l, p in zip(labels, predictions):
-        if abs(l - p) < 1e-5:
-            loss = loss + 1
-    loss = loss / len(labels)
-
-    return loss
-
-
-def cost(weights, features, labels):
-    """Cost (error) function to be minimized."""
-
-    predictions = [variational_classifier(weights, angles=f) for f in features]
-
-    return square_loss(labels, predictions)
-
-
-# load Iris data and normalise feature vectors
-data = np.loadtxt("data/iris_classes1and2_scaled.txt")
-X = data[:, 0:2]
-
-# pad the vectors to size 2^2 with constant values
-padding = 0.3 * np.ones((len(X), 1))
-X_pad = np.c_[np.c_[X, padding], np.zeros((len(X), 1)) ] 
-
-# normalize each input
-normalization = np.sqrt(np.sum(X_pad ** 2, -1))
-X_norm = (X_pad.T / normalization).T  
-
-# angles for state preparation are new features
-features = np.array([get_angles(x) for x in X_norm])   
-
-Y = data[:, -1]
-
-# split into training and validation set
-np.random.seed(0)
-num_data = len(Y)
-num_train = int(0.75 * num_data)
-index = np.random.permutation(range(num_data))
-feats_train = features[index[:num_train]]
-Y_train = Y[index[:num_train]]
-feats_val = features[index[num_train:]]
-Y_val = Y[index[num_train:]]
-
-# initialize weight layers
-num_qubits = 2
-num_layers = 6
-var_init = (0.01 * np.random.randn(num_layers, num_qubits, 3), 0.0)
-
-opt = NesterovMomentumOptimizer(0.01)
-batch_size = 5
-
-# train the variational classifier
-var = var_init
-for it in range(60):
-
-    # Update the weights by one optimizer step
-    batch_index = np.random.randint(0, num_train, (batch_size, ))
-    feats_train_batch = feats_train[batch_index]
-    Y_train_batch = Y_train[batch_index]
-    var = opt.step(lambda v: cost(v, feats_train_batch, Y_train_batch), var)
-
-    # Compute predictions on train and validation set
-    predictions_train = [np.sign(variational_classifier(var, angles=f)) for f in feats_train]
-    predictions_val = [np.sign(variational_classifier(var, angles=f)) for f in feats_val]
-
-    # Compute accuracy on train and validation set
-    acc_train = accuracy(Y_train, predictions_train)
-    acc_val = accuracy(Y_val, predictions_val)
-
-    print("Iter: {:5d} | Cost: {:0.7f} | Acc train: {:0.7f} | Acc validation: {:0.7f} "
-          "".format(it+1, cost(var, features, Y), acc_train, acc_val))
