+++ examples/figures/bloch.png
