+++ examples/figures/gaussian_transformation.png
