+++ examples/figures/hybrid_graph.png
