+++ examples/figures/photon_redirection.png
