+++ examples/pennylane_gaussian_transformation.py
@@ -16,7 +16,7 @@ The quantum circuit
 For this basic tutorial, we will consider a special subset of CV operations:
 the *Gaussian transformations*. We work with the following simple Gaussian circuit:
 
-.. figure:: ../tutorials/figures/gaussian_transformation.svg
+.. figure:: ../../examples/figures/gaussian_transformation.svg
     :align: center
     :width: 40%
     :target: javascript:void(0);
