+++ examples/pennylane_qubit_rotation.py
@@ -17,7 +17,7 @@ The quantum circuit
 
 In the qubit rotation example, we wish to implement the following quantum circuit:
 
-.. figure:: ../tutorials/figures/rotation_circuit.png
+.. figure:: ../../examples/figures/rotation_circuit.png
     :align: center
     :width: 40%
     :target: javascript:void(0);
@@ -297,7 +297,7 @@ print(dcircuit(0.54, 0.12))
 # In other words, the optimization procedure will find the weights
 # :math:`\phi_1` and :math:`\phi_2` that result in the following rotation on the Bloch sphere:
 #
-# .. figure:: ../tutorials/figures/bloch.png
+# .. figure:: ../../examples/figures/bloch.png
 #     :align: center
 #     :width: 70%
 #     :target: javascript:void(0);
