+++ examples/pennylane_skip_plugins_hybrid.py
@@ -31,7 +31,7 @@ A non-Gaussian circuit
 We first consider a photonic circuit which is similar in spirit to the
 :ref:`qubit rotation <qubit_rotation>` circuit:
 
-.. figure:: ../tutorials/figures/photon_redirection.png
+.. figure:: ../../examples/figures/photon_redirection.png
     :align: center
     :width: 30%
     :target: javascript:void(0);
@@ -297,7 +297,7 @@ def squared_difference(x, y):
 # our three subcomponents. Here, we wish to
 # perform the following hybrid quantum-classical optimization:
 #
-# .. figure:: ../tutorials/figures/hybrid_graph.png
+# .. figure:: ../../examples/figures/hybrid_graph.png
 #     :align: center
 #     :width: 70%
 #     :target: javascript:void(0);
